Move the monitor and eventhandler modules to a monitoring package.
