+++ walkingpadfitbit/domain/monitoring/eventhandler.py
